Optionally install plugins on startup (#268)
Add support for an environment variable (INSTALL_PLUGINS) that allows
installing a set of plugins on image startup. The format of the variable
is a set of comma-separated strings of the form [plugin-id]:[version].
Example:
INSTALL_PLUGINS=groovy:1.30,github-pr-comment-build:1.1

When used with persistent storage, plugin download/installation would
only happen the very first time the image is run.