Add .NET Core 2.0 kube-slave config.
The .NET Core 2.0 jenkins slave image will be delivered to
the Red Hat Container Registry and is maintained upstream
separately in this repository:
https://github.com/redhat-developer/dotnet-jenkins-slave

This patch adds relevant kubernetes slave configuration for
the .NET Core 2.0 jenkins slave image so that jobs can be
scheduled on pods with .NET Core 2.0 installed. Jobs need
to be configured - via restrictions - to only run on nodes
with label expression "dotnet-20".