Upgrade Node.js slave to Node.js 8
This is mainly just an observation that the current Node.js 4 that's
referenced in these Dockerfiles is quite old at this point. While it's
still maintained, it's in the 'Maintenance' phase of its LTS cycle,
and will be EOL in April 2018:

https://github.com/nodejs/Release#lts-schedule

I was originally going to propose Node.js 6 here, but given that
Node.js 8 is the most recent active LTS, this proposes to go straight
to that.