Add JENKINS_RELEASE var, used for label and rpm
This is a proposal based on discussion in #481. It allows you to have
one place to update the version (the new env var), and it'll be used
for both the label and the RPM installation.

The idea is that one could query the Jenkins version without having to
pull the image. I might be completely misunderstanding what's desired
in #481, but I suggested that one could use skopeo to check a label
rather than pulling the image and running it, so I figured I'd give
this example of how the label could exist, and still only have one
place to update.