Amend arguments if none passed to run-jnlp-client
As of version 1.6.0 of the Jenkins Kubernetes plugin,
args are no longer passed to jnlp container if not present in containerTemplate

See https://github.com/jenkinsci/kubernetes-plugin/pull/315

This is made worse if using Declarative pipeline and specifying yaml for the podTemplate.
since ${computer.*} is not resolvable.

To workaround this, we use JENKINS_SECRET and JENKINS_NAME for launcher arguments

See https://gist.github.com/scoheb/f29d263c240e8880648b920240635ed6 for an example Jenkinsfile